# src/internal/models/games_genres.go
package models

type GameGenres string

const (
	GameGenreUndefined GameGenres = "undefined"
	GameGenreStrategy  GameGenres = "strategy"
	GameGenreTableTop  GameGenres = "table top"
	GameGenreRpg       GameGenres = "RPG"
)